chore: children 사용
